+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/IconTimeBasedWeatherValueFactory.java
@@ -33,8 +33,8 @@ public class IconTimeBasedWeatherValueFactory extends TimeBasedWeatherValueFacto
   private static final String TEMPERATURE = "t2m";
   private static final String WIND_VELOCITY_U = "u131m";
   private static final String WIND_VELOCITY_V = "v131m";
-  private static final String GROUND_TEMPERATURE_VALUE_ONE = "t_v1";
-  private static final String GROUND_TEMPERATURE_VALUE_TWO = "t_v2";
+  private static final String GROUND_TEMPERATURE_LEVEL_1 = "tg_1";
+  private static final String GROUND_TEMPERATURE_LEVEL_2 = "tg_2";
 
   public IconTimeBasedWeatherValueFactory() {
     super();
@@ -55,8 +55,8 @@ public class IconTimeBasedWeatherValueFactory extends TimeBasedWeatherValueFacto
             "albrad",
             "asobs",
             "aswdifuS",
-            "t_v1",
-            "t_v2",
+            "tg_1",
+            "tg_2",
             "u10m",
             "u20m",
             "u216m",
@@ -91,10 +91,10 @@ public class IconTimeBasedWeatherValueFactory extends TimeBasedWeatherValueFacto
         data.getQuantity(TEMPERATURE, Units.KELVIN).to(StandardUnits.TEMPERATURE);
     ComparableQuantity<Angle> windDirection = getWindDirection(data);
     ComparableQuantity<Speed> windVelocity = getWindVelocity(data);
-    Optional<ComparableQuantity<Temperature>> groundTemperatureValueOne =
-        data.getQuantityOptional(GROUND_TEMPERATURE_VALUE_ONE, Units.KELVIN);
-    Optional<ComparableQuantity<Temperature>> groundTemperatureValueTwo =
-        data.getQuantityOptional(GROUND_TEMPERATURE_VALUE_TWO, Units.KELVIN);
+    Optional<ComparableQuantity<Temperature>> groundTemperatureLevel1 =
+        data.getQuantityOptional(GROUND_TEMPERATURE_LEVEL_1, Units.KELVIN);
+    Optional<ComparableQuantity<Temperature>> groundTemperatureLevl2 =
+        data.getQuantityOptional(GROUND_TEMPERATURE_LEVEL_2, Units.KELVIN);
 
     WeatherValue weatherValue =
         new WeatherValue(
@@ -104,8 +104,8 @@ public class IconTimeBasedWeatherValueFactory extends TimeBasedWeatherValueFacto
             temperature,
             windDirection,
             windVelocity,
-            groundTemperatureValueOne,
-            groundTemperatureValueTwo);
+            groundTemperatureLevel1,
+            groundTemperatureLevl2);
     return new TimeBasedValue<>(time, weatherValue);
   }